Fix compilation errors and add magnetosphere, albedo, and planet presets
Compilation Fixes:
- Fixed GetMeteorologicalData() -> GetMeteorology() in ForestFireManager
- Fixed double to float cast in DisasterManager earthquake deformation
- Removed non-existent Gabbro property from GeologicalSimulator
- Fixed ManualPlantingTool constructor argument order
- Fixed ManualPlantingTool.IsActive property name (was IsVisible)
- Fixed ManualPlantingTool.Update() parameter order

Magnetosphere & Cosmic Rays System (MagnetosphereSimulator.cs):
- Simulates planetary magnetic field strength (dynamo based on core temperature)
- Cosmic ray deflection by magnetosphere
- Solar wind protection (stronger protection with active magnetosphere)
- Radiation levels affected by: latitude, altitude, atmosphere, magnetic field
- Polar auroras when solar activity is high
- Magnetic field reversals (like Earth's historical events)
- Solar storm triggering capability
- Life damage from high radiation exposure
- Mars-like scenario (no magnetosphere) increases surface radiation 3x

Surface Albedo (Reflection) System (ClimateSimulator.cs):
- Realistic albedo values for different surface types:
  * Ice/Snow: 0.85 (85% reflection, stays very cold)
  * Desert sand: 0.35 (35% reflection, bright)
  * Grassland: 0.23 (23% reflection)
  * Forest: 0.17 (17% reflection, dark, absorbs heat)
  * Ocean: 0.06-0.08 (6-8% reflection, very dark)
  * Mountains/rock: 0.15 (15% reflection)
  * Cities: 0.20 (20% reflection, concrete/asphalt)
- Albedo reduces absorbed solar energy, creating ice-albedo feedback loops
- Affects global temperature dynamics

Planet Presets (MapOptionsUI.cs):
- Earth Preset (F6): 29% land, 71% water, moderate mountains, realistic
- Mars Preset (F7): 100% land (dry), high mountains, low valleys
- Water World Preset (F8): 90% water, small islands, smooth terrain
- Desert World Preset (F9): 85% land (Dune-like), fine sand detail
- All presets update map preview in real-time

Integration:
- Magnetosphere system fully integrated into game loop
- Albedo calculations applied to temperature simulation
- Preset system integrated with keyboard controls (F6-F9)
- All systems respect save/load operations

+++ DisasterManager.cs
@@ -189,7 +189,7 @@ public class DisasterManager
                 float effect = (1.0f - distance / radius) * magnitude;
 
                 // Ground deformation
-                target.Elevation += (_random.NextDouble() - 0.5) * 0.05f * effect;
+                target.Elevation += (float)(_random.NextDouble() - 0.5) * 0.05f * effect;
 
                 // Building damage (affects civilizations)
                 target.Biomass *= (1.0f - 0.3f * effect);
